- Segment some of the processor logic so it can be overriden - Adjust the stream cursor, simplified for implementation
